requirements.txt
pandas
numpy
prettytable
unidecode
mysql.connector